+++ lib/src/models/models.dart
@@ -1,2 +1,3 @@
 export 'breakpoints.dart';
+export 'grid_delegate.dart';
 export 'rtl_options.dart';